requirements.txt
langchain-cohere==0.0.1
langchain-huggingface==0.0.1
langchain-groq==0.0.1
langchain-qdrant==0.0.1
qdrant-client==1.3.0
python-dotenv==1.0.0
langchain streamlit
langchain
langchain-core
langchain-community
langserve[all]
langsmith